+++ src/input/input_ext-dec.c
@@ -2,7 +2,7 @@
  * input_ext-dec.c: services to the decoders
  *****************************************************************************
  * Copyright (C) 1998-2001 VideoLAN
- * $Id: input_ext-dec.c,v 1.41 2002/11/11 14:39:12 sam Exp $
+ * $Id: input_ext-dec.c,v 1.42 2002/11/13 20:51:05 sam Exp $
  *
  * Authors: Christophe Massiot <massiot@via.ecp.fr>
  *
@@ -203,7 +203,7 @@ void BitstreamNextDataPacket( bit_stream_t * p_bit_stream )
 u32 UnalignedShowBits( bit_stream_t * p_bit_stream, unsigned int i_bits )
 {
     /* We just fill in the bit buffer. */
-    while( p_bit_stream->fifo.i_available < i_bits )
+    while( (unsigned int)p_bit_stream->fifo.i_available < i_bits )
     {
         if( p_bit_stream->p_byte < p_bit_stream->p_end )
         {
@@ -221,7 +221,7 @@ u32 UnalignedShowBits( bit_stream_t * p_bit_stream, unsigned int i_bits )
                 /* We are not aligned anymore. */
                 if( ((ptrdiff_t)p_bit_stream->p_byte
                                     & (sizeof(WORD_TYPE) - 1)) * 8
-                        < p_bit_stream->fifo.i_available )
+                        < (unsigned int)p_bit_stream->fifo.i_available )
                 {
                     /* We are not aligned, and won't be. Copy the first word
                      * of the packet in a temporary buffer, and we'll see
